Moved getTestCredentials method to TestCredentials class
