+++ app/src/androidTest/java/com/microsoft/graph/snippets/TestCredentials.java
@@ -4,8 +4,34 @@
  */
 package com.microsoft.graph.snippets;
 
+import android.os.Environment;
+
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
+
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.FileReader;
+
 public class TestCredentials {
-    public String testClientId;
-    public String testUsername;
-    public String testPassword;
+    private static final String TEST_ARTIFACT_LOCATION = "local/testConfig.json";
+    private static final String CLIENT_ID_TEST_ARTIFACT = "test_client_id";
+    private static final String USERNAME_TEST_ARTIFACT = "test_username";
+    private static final String PASSWORD_TEST_ARTIFACT = "test_password";
+
+    public String clientId;
+    public String username;
+    public String password;
+
+    public static TestCredentials getTestCredentials() throws FileNotFoundException {
+        TestCredentials testCredentials = new TestCredentials();
+        File testConfigFile = new File(Environment.getDataDirectory(), TEST_ARTIFACT_LOCATION);
+        JsonObject testConfig = new JsonParser().parse(new FileReader(testConfigFile)).getAsJsonObject();
+
+        testCredentials.clientId = testConfig.get(CLIENT_ID_TEST_ARTIFACT).getAsString();
+        testCredentials.username = testConfig.get(USERNAME_TEST_ARTIFACT).getAsString();
+        testCredentials.password = testConfig.get(PASSWORD_TEST_ARTIFACT).getAsString();
+
+        return testCredentials;
+    }
 }
